pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
